Change function name to sendEmailAfterBondingPeriodHasEndedNotificationToDelegators
